+++ layouts/index.ts
@@ -1,0 +1,13 @@
+import {
+  SimpleLayout,
+  DocsLayout,
+  UnstyledLayout,
+  BlogLayout,
+} from "@portaljs/core";
+
+export default {
+  simple: SimpleLayout,
+  docs: DocsLayout,
+  unstyled: UnstyledLayout,
+  blog: BlogLayout,
+};
