+++ package.json
@@ -12,38 +12,28 @@
     "generate": "cross-env NODE_OPTIONS=\"--experimental-json-modules --experimental-modules\" node -r esbuild-register ./scripts/prebuild.mjs"
   },
   "dependencies": {
-    "@docsearch/react": "^3.2.0",
-    "@headlessui/react": "^1.6.6",
-    "@splidejs/react-splide": "^0.7.12",
-    "clsx": "^1.2.1",
-    "crypto-js": "^4.1.1",
-    "d3": "^7.6.1",
-    "d3-ternary": "^2.0.14",
-    "date-fns": "^2.29.2",
-    "fuse.js": "^6.6.2",
     "@heroicons/react": "^1.0.6",
+    "@portaljs/core": "^1.0.5",
     "@portaljs/remark-callouts": "^1.0.5",
     "@portaljs/remark-embed": "^1.0.4",
     "@portaljs/remark-wiki-link": "^1.0.4",
     "@silvenon/remark-smartypants": "^2.0.0",
+    "@splidejs/react-splide": "^0.7.12",
     "@tailwindcss/typography": "^0.5.9",
     "@types/node": "18.14.2",
     "@types/react": "18.0.28",
     "@types/react-dom": "18.0.11",
+    "date-fns": "^2.29.3",
     "gray-matter": "^4.0.3",
-    "itemsjs": "^2.1.15",
-    "kbar": "^0.1.0-beta.37",
-    "next-themes": "^0.2.1",
     "hastscript": "^7.2.0",
+    "kbar": "^0.1.0-beta.37",
     "mddb": "^0.1.9",
     "mdx-mermaid": "2.0.0-rc7",
     "next": "13.3.0",
     "next-mdx-remote": "^4.4.1",
     "next-seo": "^5.15.0",
     "react": "^18.2.0",
-    "react-aria": "^3.20.0",
     "react-dom": "^18.2.0",
-    "react-stately": "^3.18.0",
     "rehype-autolink-headings": "^6.1.1",
     "rehype-katex": "^6.0.2",
     "rehype-prism-plus": "^1.5.1",
@@ -52,7 +42,6 @@
     "remark-math": "^5.1.1",
     "remark-smartypants": "^2.0.0",
     "remark-toc": "^8.0.1",
-    "sort-array": "^4.1.5",
     "strip-markdown": "^5.0.0",
     "typed.js": "^2.0.12",
     "typescript": "^4.9.5"
