+++ styles/global.css
@@ -1,31 +1,12 @@
-@import "remark-callouts/styles.css";
-@import '@splidejs/react-splide/css/core';
+@import "@portaljs/remark-callouts/styles.css";
 
-@import url("https://fonts.googleapis.com/css2?family=EB+Garamond&display=swap");
-
-/* Fonts */
-@font-face {
-  font-family: "Restora";
-  src: url("/fonts/Restora.ttf");
-  font-style: normal;
-  font-display: swap;
-}
-@font-face {
-  font-family: "Restora";
-  src: url("/fonts/RestoraBold.ttf");
-  font-style: bold;
-  font-display: swap;
-}
-@font-face {
-  font-family: "Restora";
-  src: url("/fonts/RestoraExtraLight.otf");
-  font-style: light;
-  font-display: swap;
+.nav-logo {
+  height: 2.25rem;
 }
 
 /* mathjax */
-.math-inline {
-  display: flex;
+.math-inline > mjx-container > svg {
+  display: inline;
   align-items: center;
 }
 
@@ -34,19 +15,39 @@ html {
   scroll-behavior: smooth !important;
 }
 
-/* hide anchor link with # icon  */
-.unstyled :is(h1, h2, h3, h4, h5, h6) > a {
+/* tooltip fade-out clip */
+.tooltip-body::after {
+  content: "";
+  position: absolute;
+  right: 0;
+  top: 3.6rem; /* multiple of $line-height used on the tooltip body (defined in tooltipBodyStyle) */
+  height: 1.2rem; /* ($top + $height)/$line-height is the number of lines we want to clip tooltip text at*/
+  width: 10rem;
+  background: linear-gradient(
+    to right,
+    rgba(255, 255, 255, 0),
+    rgba(255, 255, 255, 1) 100%
+  );
+}
+
+/* Heading links with # icons on hover */
+.unstyled .heading-link {
   visibility: hidden;
 }
 
-.docs :is(h1, h2, h3, h4, h5, h6) {
+#mdxpage iframe {
+  height: 80vh;
+}
+
+/* :not(.unstyled) :is(h2, h3, h4, h5, h6):not(.blogitem-title) { */
+#mdxpage :is(h2, h3, h4, h5, h6) {
   margin-left: -2rem !important;
   padding-left: 2rem !important;
   scroll-margin-top: 4.5rem;
   position: relative;
 }
 
-.docs :is(h1, h2, h3, h4, h5, h6) > a {
+#mdxpage .heading-link {
   padding: 1px;
   position: absolute;
   left: 0;
@@ -59,22 +60,22 @@ html {
   transition: opacity 0.2s;
 }
 
-.light .docs :is(h1, h2, h3, h4, h5, h6) > a {
+#mdx-page .light .heading-link {
   /* border: 1px solid #ab2b65; */
   /* background: none; */
   background: #e2e8f0;
 }
 
-.docs :is(h1, h2, h3, h4, h5, h6):hover > a {
+#mdxpage :is(h2, h3, h4, h5, h6):hover .heading-link {
   opacity: 100;
 }
 
-.docs :is(h1, h2, h3, h4, h5, h6) > a > svg {
+#mdxpage .heading-link svg {
   transform: scale(0.75);
 }
 
 @media screen and (max-width: 640px) {
-  .docs :is(h1, h2, h3, h4, h5, h6) > a {
+  #mdxpage .heading-link {
     visibility: hidden;
   }
 }
